Restructure libraries folder and impors to work with Arduino IDE 1.8.*
