+++ src/libraries/PacketSerial/src/Encoding/COBS.h
