+++ src/libraries/SerialCommand/LICENSE
