+++ src/libraries/SerialCommand/examples/SerialCommandExampleSoftwareSerial/SerialCommandExampleSoftwareSerial.ino
