+++ src/libraries/SerialCommand/examples/SoftwareSerialCheck/SoftwareSerialCheck.ino
